firestore-mgr/requirements.txt
google-cloud-logging==3.0.0
pytest-mock==3.7.0
